examples/weather-ai-app/pyproject.toml
[project]
name = "weather-ai-app"
version = "1.0.0"
description = "Weather AI App with Natural Language Queries"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.12",
    "uvicorn>=0.34.3",
    "pydantic>=2.11.5",
    "python-dotenv>=1.1.0",
    "httpx>=0.28.1",
    "openai>=1.58.1",
]

[dependency-groups]
dev = [
    "ruff>=0.11.13",
    "pytest>=8.0.0",
]

[tool.ruff]
exclude = [".venv"]

[tool.setuptools]
packages = ["app"]